Preliminary Data.Functor clean up
