+++ src/Control/Monad/Linear/Internal.hs
@@ -16,6 +16,7 @@ import Data.Functor.Identity
 import Data.Functor.Sum
 import Data.Functor.Compose
 import Data.Monoid.Linear hiding (Sum)
+import qualified Data.Applicative.Linear as Data
 import qualified Data.Functor.Linear.Internal as Data
 import qualified Control.Monad.Trans.Reader as NonLinear
 import qualified Control.Monad.Trans.State.Strict as NonLinear
