+++ src/Control/Optics/Linear/Internal.hs
@@ -43,6 +43,7 @@ import Data.Bifunctor.Linear (SymmetricMonoidal)
 import Data.Profunctor.Linear
 import Data.Functor.Compose hiding (getCompose)
 import Data.Functor.Linear
+import Data.Traversable.Linear
 import qualified Data.Profunctor.Kleisli.Linear as Linear
 import Data.Void
 import GHC.Exts (FUN)
