+++ src/Data/Profunctor/Kleisli/Linear.hs
@@ -44,6 +44,7 @@ import Prelude.Linear (Either(..), either)
 import Prelude.Linear.Internal
 import qualified Control.Monad.Linear as Control
 import qualified Data.Functor.Linear as Data
+import qualified Data.Applicative.Linear as Data
 
 -- Ideally, there would only be one Kleisli arrow, parametrised by
 -- a multiplicity parameter:
