+++ src/Streaming/Internal/Type.hs
@@ -16,6 +16,7 @@ module Streaming.Internal.Type
   ) where
 
 import qualified Data.Functor.Linear as Data
+import qualified Data.Applicative.Linear as Data
 import qualified Control.Monad.Linear as Control
 import qualified Prelude.Linear as Linear
 import Prelude.Linear (($), (.))
